refactor(parsers): update files structure
