+++ src/parsers/vast/vast.test.js
