+++ src/parsers/vmap/vmap.js
@@ -1,6 +1,7 @@
 import { Log } from '@clappr/core'
-import AdBreak from '../entities/ad-break'
-import { hmsToMilliseconds, xml2json } from '../utils'
+import { hmsToMilliseconds } from '../../utils/str-to-time'
+import AdBreak from '../../entities/ad-break'
+import xml2json from '../../converts/xml2json'
 
 export default class VMAPManager {
   /**
